cleaned up server.js. updated npm script to run server
