+++ package.json
@@ -4,24 +4,23 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "webpack --mode production",
-    "prod-server": "npm run build && node src/server.js",
-    "start": "webpack-dev-server --config ./webpack.config.js --mode development",
+    "build": "webpack --mode production --config ./webpack-back.config.js && webpack --mode production --config webpack-front.config.js",
+    "run-production": "npm run build && node dist/bundle-back.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/cli": "^7.0.0",
     "@babel/core": "^7.2.2",
     "@babel/plugin-proposal-class-properties": "^7.3.0",
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
-    "@babel/cli": "^7.0.0",
     "babel-loader": "^8.0.5",
+    "babel-preset-env": "^1.7.0",
     "file-loader": "^3.0.1",
     "raw-loader": "^1.0.0",
-    "react-hot-loader": "^4.6.5",
     "url-loader": "^1.1.2",
     "webpack": "^4.29.1",
     "webpack-cli": "^3.2.3",
