+++ src/client/index.js
@@ -5,7 +5,7 @@ import App from './App.js'
 
 //TODO need to set up custom express routes ... possible us nextjs
 //TODO add readme for github and a TODO doc
-ReactDOM.render(
+ReactDOM.hydrate(
 	<Router>
 		<App />
 	</Router>,
